Turn metrics updating into a system
